libebpf/bin/tools/requirements.txt
# pypi version of parcon does not support python3
git+https://github.com/javawizard/parcon
nose ~= 1.3.1
pyelftools ~= 0.23